Fix consolidation name clash for an overlapping edge case

+++ nautilus_trader/persistence/catalog/parquet.py
@@ -654,7 +654,8 @@ class ParquetDataCatalog(BaseDataCatalog):
         pq.write_table(combined_table, where=new_file)
 
         for file in file_list:
-            self.fs.rm(file)
+            if file != new_file:
+                self.fs.rm(file)
 
     def consolidate_catalog_by_period(
         self,
